(Bonus Build Fail) Adding targets for iOS
- We first start with declaring a target for the `transmission` and
`transmission-test` libraries. These will set up the library usage for
ios targets in the future. For the sample app, we similarly need to
define the targets we want to show this app to. So we add ios related
targets.
- For the sample, we want to use the whole sample code in iOS, so we
bundle the app code into a Framework by creating a framework for each
native type.

- However, build now fails. Since we added new ios targets, our
  commonMain code is not truly Multiplatform. We need to replace java
  specific code